# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WA-AKG API Documentation",
    "version": "1.2.0",
    "description": "\n# WhatsApp AI Gateway - Complete API Reference\n\nProfessional WhatsApp Gateway with **64 API endpoints** for complete WhatsApp automation.\n\n## 🔐 Authentication\nAll endpoints require authentication via:\n1. **API Key Header**: `X-API-Key: your-api-key`\n2. **Session Cookie**: `next-auth.session-token` (automatically sent by browser)\n\n## 📋 Common Parameters\n- **sessionId**: Unique session identifier (e.g., \"mysession-01\")\n- **jid**: WhatsApp JID format:\n  - Personal: `628123456789@s.whatsapp.net`\n  - Group: `120363123456789@g.us`\n\n## 📊 Rate Limits\n- Phone check: Max 50 numbers per request\n- Broadcast: 10-20s random delay between messages\n- Message history: Max 100 messages\n                "
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "API Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API Key for authentication"
      },
      "SessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token",
        "description": "Session cookie (browser only)"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "status": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Error occurred"
          },
          "error": {
            "type": "string",
            "example": "Detailed error info"
          }
        }
      },
...
      "Contact": {
        "type": "object",
        "properties": {
          "jid": {
            "type": "string",
            "example": "628123456789@s.whatsapp.net"
          },
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "notify": {
            "type": "string"
          },
          "profilePic": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "Group": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "subject": {
            "type": "string"
          },
          "desc": {
            "type": "string"
          },
          "owner": {
            "type": "string"
          },
          "size": {
            "type": "number"
          },
          "isCommunity": {
            "type": "boolean",
            "description": "Whether this group is a WhatsApp Community announcement group"
          },
          "linkedParentJid": {
            "type": "string",
            "nullable": true,
            "description": "JID of the parent community if this is a sub-group"
          },
          "participants": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "admin": {
                  "type": "string",
                  "nullable": true
                }
              }
            }
          }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Unauthorized - Invalid or missing API key",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "status": false,
              "message": "Unauthorized",
              "error": "Unauthorized"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden - Access denied",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "status": false,
              "message": "Forbidden - Cannot access this session",
              "error": "Forbidden - Cannot access this session"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "status": false,
              "message": "Session not found",
              "error": "Session not found"
            }
          }
        }
      },
      "SessionNotReady": {
        "description": "Session not connected or ready",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "status": false,
              "message": "Session not ready",
              "error": "Session not ready"
            }
          }
        }
      },
...
      "ServerError": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "status": false,
              "message": "Internal Server Error",
              "error": "Internal Server Error"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "SessionAuth": []
    }
  ],
  "paths": {
...
    "/groups/{sessionId}": {
      "get": {
        "tags": [
          "Groups"
        ],
        "summary": "List all groups",
        "description": "Get all groups associated with the session",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of groups",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Group"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "description": "Failed to fetch groups"
          }
        }
      }
    },
...
    "/groups/{sessionId}/create": {
      "post": {
        "tags": [
          "Groups"
        ],
        "summary": "Create new group",
        "description": "Create a new group with specified participants",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "subject",
                  "participants"
                ],
                "properties": {
                  "subject": {
                    "type": "string",
                    "maxLength": 100,
                    "example": "VIP Customers"
                  },
                  "participants": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": [
                      "628123456789@s.whatsapp.net"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Group created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "group": {
                      "$ref": "#/components/schemas/Group"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "description": "Failed to create group"
          },
          "503": {
            "$ref": "#/components/responses/SessionNotReady"
          }
        }
      }
    },
...
    "/groups/{sessionId}/{jid}/invite": {
      "get": {
        "tags": [
          "Groups"
        ],
        "summary": "Get invite code",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "jid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invite code",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "inviteCode": {
                      "type": "string"
                    },
                    "inviteUrl": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "description": "Failed to get code"
          }
        }
      },
      "put": {
        "tags": [
          "Groups"
        ],
        "summary": "Revoke invite code",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "jid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invite revoked",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string"
                    },
                    "newInviteCode": {
                      "type": "string"
                    },
                    "inviteUrl": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "description": "Failed to revoke code"
          }
        }
      }
    },
...
    "/contacts": {
      "get": {
        "tags": [
          "Contacts"
        ],
        "summary": "List contacts (DEPRECATED)",
        "description": "**DEPRECATED:** Use GET /contacts/{sessionId} instead. This endpoint will be removed in a future version.",
        "deprecated": true,
        "parameters": [
          {
            "name": "sessionId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Search by name, notify, jid"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated contacts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Contact"
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "integer"
                        },
                        "page": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "totalPages": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/groups/{jid}": {
      "get": {
        "tags": [
          "Groups"
        ],
        "summary": "Get group details (DEPRECATED)",
        "description": "**DEPRECATED:** Use GET /groups/{sessionId}/{jid} instead. This endpoint will be removed in a future version.",
        "deprecated": true,
        "parameters": [
          {
            "name": "jid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sessionId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Group details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "subject": {
                      "type": "string"
                    },
                    "subjectOwner": {
                      "type": "string"
                    },
                    "subjectTime": {
                      "type": "integer"
                    },
                    "size": {
                      "type": "integer"
                    },
                    "creation": {
                      "type": "integer"
                    },
                    "owner": {
                      "type": "string"
                    },
                    "desc": {
                      "type": "string"
                    },
                    "descId": {
                      "type": "string"
                    },
                    "restrict": {
                      "type": "boolean"
                    },
                    "announce": {
                      "type": "boolean"
                    },
                    "participants": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "admin": {
                            "type": "string",
                            "nullable": true
                          }
                        }
                      }
                    },
                    "ephemeralDuration": {
                      "type": "integer"
                    },
                    "inviteCode": {
                      "type": "string"
                    },
                    "pictureUrl": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "description": "Group not found"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  },
...
}